fix: wait for Android exposure capture readiness
